+++ CHANGELOG.md
@@ -9,7 +9,7 @@
  - Use all `IListMatrixSet` options in `getTileGrid` to create `WMTSTileGrid`.
 
 * **@dlr-eoc/services-layers:**
- - More options to create `WMTSTileGrid` on `IListMatrixSet`. The new optional options are `extent?: TGeoExtent`, `origin?: number[]`, `origins?: [number, number][]` and `tileSizes?: number[]`.
+ - More options to create `WMTSTileGrid` on `IListMatrixSet`. The new optional options are `extent?: TGeoExtent`, `origin?: number[]`, `origins?: [number, number][]`, `tileSizes?: number[]` and `sizes?: [number, number][];`.
 
  
 ### Bug Fixes